重构trace聚合根实体结构修改.

+++ traffic-service/api/src/main/java/com/oc/hawk/traffic/entrypoint/api/command/UploadTraceInfoCommand.java
@@ -15,7 +15,7 @@ public class UploadTraceInfoCommand {
     private String dstWorkload;
     private String dstNamespace;
     private Long timestamp;
-    private Integer latency;
+    private Long latency;
     private String requestId;
     private String protocol;
     private String method;
